+++ CHANGES.rst
@@ -34,6 +34,8 @@ Fixed:
 - Limit displayed journals to current document pool. #4515
   [tlotze]
 
+- Catch SQLAlchemy error when querying transfers. #4758
+  [tlotze]
 
 Changed:
 
@@ -61,7 +63,7 @@ Fixed:
 
 - Missing translations in Chronologie #3708
   [slindner]
- 
+
 - Don't log login times to member properties to avoid DB hotspot. #4325
   [tlotze]